requirements.txt
absl-py==1.4.0
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.1.0
astunparse==1.6.3
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.3.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.6
comtypes==1.1.14
contourpy==1.0.5
cryptography==39.0.0
curses-2048==1.4
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.5
Django==4.1
django-bootstrap-v5==1.0.11
django-bootstrap4==22.2
django-cors-headers==3.13.0
django-filter==22.1
django-location-field==2.1.0
django-request==1.6.1
django-rest-knox==4.2.0
django-versatileimagefield==3.0
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.2
drf-flex-fields==1.0.0
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.7.1
flatbuffers==23.3.3
fonttools==4.37.4
frozendict==2.3.6
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.8.0
html5lib==1.1
idna==3.4
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
jax==0.4.8
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.10
jsonschema==4.17.0
jupyter-server==1.23.3
jupyter_client==7.4.7
jupyter_core==5.0.0
jupyterlab==3.5.0
jupyterlab-pygments==0.2.2
jupyterlab_server==2.16.3
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
ml-dtypes==0.1.0
multitasking==0.0.11
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
prometheus-client==0.15.0
prompt-toolkit==3.0.33
protobuf==4.22.4
psutil==5.9.4
psycopg2==2.9.3
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
pypiwin32==223
pypng==0.20220715.0
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
python-magic-bin==0.4.14
pyttsx3==2.90
pytz==2023.3
pywin32==305
pywinpty==2.0.9
pyzmq==24.0.1
qrcode==7.4.2
requests==2.28.1
requests-oauthlib==1.3.1
response==0.5.0
rsa==4.9
scikit-learn==1.2.0
scipy==1.9.3
seaborn==0.12.1
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.6.1
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.0
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
traitlets==5.5.0
typing_extensions==4.5.0
tzdata==2022.1
urllib3==1.26.12
virtualenv==20.16.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.3.3
windows-curses==2.3.0
wrapt==1.14.1
yfinance==0.2.14
